fix: 풀리퀘스트 내용 반영
